Fixed missing tilde expansion
The interactive setup threw an error if the user tried to use ~ as home

Tilde is often used to refer to the current users HOME directory but is
actually only a shell feature and not reflected in unix itself. The
shell is automatically expanding the tilde to the currect home
directory.
Getting the wallet and ledger directory via the builtin read command
bypasses the tilde expansion and results in a path containing the
literal '~'

Signed-off-by: Cristian Kubis <cristian.kubis@tsunix.de>